Add constraints to dragging behavior
- For all diagrams: constrain drag position to stay within the diagram
- For graph diagrams: constrain circles to not overlap

The drag behavior isn't always smooth because there's no collision
response. It stops the circle from moving but sometimes we could've
moved it partially or have the circle slide around another circle.
That feels better but it's hard to implement, especially when there
are more than two circles, so it will have to wait.